Fix timestamp encoding to use Eagle Time oscillation counts
Changed from storing timestamps as f64 seconds to usize oscillation counts:
- Updated PatchMetadata.timestamp from f64 to usize
- Changed VsfType::e(EtType::f6) to VsfType::e(EtType::u)
- Use eagle_time::eagle_time_oscillations() instead of Unix timestamp
- Updated all test fixtures and encode/decode functions

This ensures patches use eu6 format (unsigned 64-bit oscillation counts)
instead of ef5/ef6 (float), giving proper picosecond precision.

Requires vsf >= 0.2.3 with Eagle Time display fixes.

+++ src/patch.rs
@@ -20,8 +20,8 @@ pub struct PatchMetadata {
     /// Parent patch ID (None for initial patch)
     pub parent: Option<PatchId>,
 
-    /// Creation timestamp (Eagle Time - seconds since 1969-07-20 20:17:40 UTC)
-    pub timestamp: f64,
+    /// Creation timestamp (Eagle Time - oscillation count since 1969-07-20 20:17:40 UTC)
+    pub timestamp: usize,
 
     /// Patch message
     pub message: String,
@@ -106,7 +106,7 @@ impl Patch {
     pub fn new(
         author: AuthorId,
         parent: Option<PatchId>,
-        timestamp: f64,
+        timestamp: usize,
         message: String,
         operations: Vec<FileOp>,
         build_hash: [u8; 32],
